fix: 평균 구매 가격 저장 오류 해결

+++ src/main/java/com/mock/investment/stock/domain/account/dto/HoldingStockDto.java
@@ -19,7 +19,7 @@ public class HoldingStockDto {
     private BigDecimal quantity;
 
     @PositiveOrZero
-    private BigDecimal avgPrice;
+    private Double avgPrice;
 
     public static List<HoldingStockDto> fromEntities(List<HoldingStock> holdingStocks) {
         return holdingStocks.stream()
